# backend/internal/api/routes.go
package api

import (
	"bytes"
	"io/fs"
	"net/http"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
)

func (s Server) registerRoutes(router *gin.Engine) {
	router.GET("/api/health", s.health)
	router.GET("/api/ready", s.ready)
	router.GET("/api/patch/info", s.patchInfo)
	authPublic := router.Group("/api/auth")
	authPublic.Use(SameOriginWrite())
	authPublic.GET("/status", s.authStatus)
	authPublic.POST("/register", s.registerAdmin)
	authPublic.POST("/login", s.login)
	breed := router.Group("/api/breed")
	breed.Use(SameOriginWrite())
	breed.POST("/session/exchange", s.exchangeBreedSession)
	breed.Use(s.breedSessionAuth())
	breed.GET("/me", s.breedSessionMe)
	breed.GET("/catalog", s.breedCatalog)
	breed.GET("/history", s.breedHistory)
	breed.GET("/presets", s.listBreedingPresets)
	breed.POST("/presets", s.putBreedingPreset)
	breed.DELETE("/presets/:id", s.deleteBreedingPreset)
	breed.GET("/custom-containers", s.listCustomPalContainers)
	breed.POST("/custom-containers", s.putCustomPalContainer)
	breed.DELETE("/custom-containers/:id", s.deleteCustomPalContainer)
	breed.POST("/jobs", s.breedSubmitJob)
	breed.GET("/jobs/:id", s.breedJob)
	breed.GET("/jobs/:id/result", s.breedJobResult)
	breed.POST("/jobs/:id/pause", s.breedControlJob("pause"))
	breed.POST("/jobs/:id/resume", s.breedControlJob("resume"))
	breed.POST("/jobs/:id/cancel", s.breedControlJob("cancel"))
	integration := router.Group("/api/integrations/astrbot")
	integration.Use(s.astrBotSignatureAuth())
	integration.POST("/binding-challenges", s.astrBotBindingChallenge)
	integration.POST("/quick-solves", s.astrBotQuickSolve)
	integration.POST("/server-status", s.astrBotServerStatus)
	integration.POST("/server-control", s.astrBotServerControl)
	integration.POST("/community-servers", s.astrBotCommunityServers)

	api := router.Group("/api")
	api.Use(Auth(s.cfg, s.auth), DetailedAuditMiddleware(s.store))
	s.registerSystemRoutes(api)
	s.registerServerRoutes(api)
	s.registerContentRoutes(api)
	s.registerSecurityRoutes(api)
	s.registerWorldRoutes(api)
	api.GET("/catalog", s.apiCatalog(router))
	s.registerFrontendRoutes(router)
}

func (s Server) registerSystemRoutes(api *gin.RouterGroup) {
	api.GET("/auth/me", s.authMe)
	api.POST("/auth/logout", s.logout)
	api.PUT("/auth/password", s.changePassword)
	api.GET("/auth/api-keys", s.listAPIKeys)
	api.POST("/auth/api-keys", Require(PermSecurityWrite), s.createAPIKey)
	api.DELETE("/auth/api-keys/:id", Require(PermSecurityWrite), s.revokeAPIKey)
	api.GET("/jobs", s.listJobs)
	api.GET("/jobs/:id", s.getJob)
	api.GET("/patch/update/status", s.patchUpdateStatus)
	api.POST("/patch/update/check", Require(PermServerControl), s.patchUpdateCheck)
	api.POST("/patch/update", Require(PermServerControl), s.patchUpdate)
	api.GET("/audit-logs", Require(PermAuditRead), s.listAuditLogs)
	api.GET("/system/debug", s.debugLoggingStatus)
	api.PUT("/system/debug", Require(PermConfigWrite), s.putDebugLogging)
	api.GET("/settings/network-proxy", s.getNetworkProxyConfig)
	api.PUT("/settings/network-proxy", Require(PermConfigWrite), s.putNetworkProxyConfig)
	api.POST("/settings/network-proxy/test", Require(PermConfigWrite), s.testNetworkProxy)
	api.GET("/alerts", s.listAlerts)
	api.POST("/alerts/:id/ack", Require(PermServerControl), s.ackAlert)
	api.GET("/schedules", s.listSchedules)
	api.POST("/schedules", Require(PermServerControl), s.createSchedule)
	api.PUT("/schedules/:id", Require(PermServerControl), s.updateSchedule)
	api.DELETE("/schedules/:id", Require(PermServerControl), s.deleteSchedule)
	api.POST("/schedules/:id/run", Require(PermServerControl), s.runSchedule)
}

func (s Server) registerServerRoutes(api *gin.RouterGroup) {
	api.GET("/community-servers", s.listCommunityServers)
	api.GET("/community-servers/source-status", s.communityServersSourceStatus)
	api.POST("/community-servers/refresh", Require(PermRead), s.refreshCommunityServers)
	api.GET("/server/status", s.serverStatus)
	api.GET("/server/prerequisites", s.serverPrerequisites)
	api.GET("/server/host", s.serverHost)
	api.GET("/server/runtime", s.getRuntime)
	api.PUT("/server/runtime", Require(PermConfigWrite), s.putRuntime)
	api.POST("/server/import", Require(PermConfigWrite), s.importServerDirectory)
	api.GET("/server/docker/plan", s.serverDockerPlan)
	api.POST("/server/docker/install", Require(PermServerControl), s.serverDockerInstall)
	api.GET("/server/docker/mirrors/plan", s.serverDockerMirrorsPlan)
	api.POST("/server/docker/mirrors/configure", Require(PermServerControl), s.serverDockerMirrorsConfigure)
	api.POST("/server/bootstrap", Require(PermServerControl), s.serverBootstrap)
	api.GET("/server/logs", s.serverLogs)
	api.GET("/server/world", s.serverWorld)
	api.POST("/server/world/reset", Require(PermWorldReset), s.serverWorldReset)
	api.POST("/server/install", Require(PermServerControl), s.serverInstall)
	api.POST("/server/update", Require(PermServerControl), s.serverUpdate)
	api.POST("/server/update-if-needed", Require(PermServerControl), s.serverUpdateIfNeeded)
	api.GET("/server/version", s.serverVersion)
	api.POST("/server/version/check", Require(PermServerControl), s.serverVersionCheck)
	api.POST("/server/start", Require(PermServerControl), s.serverStart)
	api.POST("/server/stop", Require(PermServerControl), s.serverStop)
	api.POST("/server/restart", Require(PermServerControl), s.serverRestart)
	api.POST("/server/safe-restart", Require(PermServerControl), s.serverSafeRestart)
	api.POST("/server/safe-stop", Require(PermServerControl), s.serverSafeStop)
	api.POST("/server/force-stop", Require(PermServerControl), s.serverForceStop)
	api.GET("/server/startup", s.getStartup)
	api.PUT("/server/startup", Require(PermConfigWrite), s.putStartup)
	api.POST("/server/initialize-config", Require(PermConfigWrite), s.initializeConfig)
	api.GET("/monitor/snapshot", s.monitorSnapshot)
	api.GET("/monitor/history", s.monitorHistory)
	api.POST("/server/backup", Require(PermBackupWrite), s.serverBackup)
	api.GET("/backups", s.listBackups)
	api.GET("/backups/webdav/config", Require(PermBackupWrite), s.getWebDAVConfig)
	api.PUT("/backups/webdav/config", Require(PermBackupWrite), s.putWebDAVConfig)
	api.POST("/backups/webdav/test", Require(PermBackupWrite), s.testWebDAVConfig)
	api.POST("/backups/:name/restore", Require(PermBackupWrite), s.restoreBackup)
	api.GET("/backups/:name/download", Require(PermBackupWrite), s.downloadBackup)
	api.DELETE("/backups/:name", Require(PermBackupWrite), s.deleteBackup)
	api.POST("/backups/:name/verify", Require(PermBackupWrite), s.verifyBackup)
	api.POST("/backups/:name/upload-webdav", Require(PermBackupWrite), s.uploadBackupToWebDAV)
	api.GET("/server/info", s.palGet("info"))
	api.GET("/server/players", s.palGet("players"))
	api.GET("/server/settings", s.palGet("settings"))
	api.GET("/server/metrics", s.serverMetrics)
	api.GET("/server/game-data", s.serverGameData)
	api.POST("/server/announce", Require(PermServerControl), s.palPost("announce"))
	api.POST("/server/save", Require(PermServerControl), s.palPost("save"))
	api.POST("/server/shutdown", Require(PermServerControl), s.palPost("shutdown"))
}

func (s Server) registerContentRoutes(api *gin.RouterGroup) {
	api.GET("/config/palworld", Require(PermConfigWrite), s.getPalworldConfig)
	api.PUT("/config/palworld", Require(PermConfigWrite), s.updatePalworldConfig)
	api.POST("/config/palworld/apply", Require(PermConfigWrite), Require(PermServerControl), s.applyPalworldConfig)
	api.GET("/config/palworld/schema", Require(PermConfigWrite), s.getPalworldConfigSchema)
	api.POST("/config/palworld/validate", Require(PermConfigWrite), s.validatePalworldConfig)
	api.GET("/mods", s.listMods)
	api.POST("/mods/local/scan", Require(PermRead), s.scanLocalMods)
	api.POST("/mods/local/findings/:id/actions", Require(PermModsWrite), s.actOnLocalModFinding)
	api.GET("/mods/workshop/status", s.workshopStatus)
	api.GET("/mods/workshop/auth/status", s.workshopAuthStatus)
	api.POST("/mods/workshop/auth/start", Require(PermSecurityWrite), s.startWorkshopAuth)
	api.POST("/mods/workshop/auth/verify", Require(PermSecurityWrite), s.verifyWorkshopAuth)
	api.DELETE("/mods/workshop/auth/credentials", Require(PermSecurityWrite), s.clearWorkshopAuth)
	api.GET("/mods/workshop/search", s.searchWorkshopMods)
	api.GET("/mods/workshop/:id", s.getWorkshopMod)
	api.POST("/mods/workshop/:id/translate", Require(PermModsWrite), s.translateWorkshopMod)
	api.POST("/mods/import/inspect", Require(PermModsWrite), s.inspectModImport)
	api.POST("/mods/import/inspect/:id/select", Require(PermModsWrite), s.selectModImportCandidate)
	api.POST("/mods/import", Require(PermModsWrite), s.startModImport)
	api.POST("/mods/upload", Require(PermModsWrite), s.uploadMod)
	api.POST("/mods/workshop", Require(PermModsWrite), s.downloadWorkshop)
	s.registerModConfigurationRoutes(api)
	api.POST("/mods/:id/enable", Require(PermModsWrite), s.enableMod)
	api.POST("/mods/:id/disable", Require(PermModsWrite), s.disableMod)
	api.DELETE("/mods/:id", Require(PermModsWrite), s.deleteMod)
	api.GET("/ai/translation/config", s.getAITranslationConfig)
	api.PUT("/ai/translation/config", Require(PermAIConfig), s.putAITranslationConfig)
	api.POST("/ai/translation/test", Require(PermAIConfig), s.testAITranslationConfig)
}

func (s Server) registerSecurityRoutes(api *gin.RouterGroup) {
	api.GET("/security/paldefender/releases", s.palDefenderReleases)
	api.GET("/security/paldefender/status", s.palDefenderStatus)
	api.POST("/security/paldefender/install", Require(PermSecurityWrite), s.palDefenderInstall)
	api.POST("/security/paldefender/update", Require(PermSecurityWrite), s.palDefenderUpdate)
	api.POST("/security/paldefender/rollback", Require(PermSecurityWrite), s.palDefenderRollback)
	api.GET("/security/paldefender/config", s.palDefenderGetConfig)
	api.PUT("/security/paldefender/config", Require(PermSecurityWrite), s.palDefenderPutConfig)
	api.POST("/security/paldefender/apply-preset", Require(PermSecurityWrite), s.palDefenderApplyPreset)
	api.POST("/security/paldefender/rest-token", Require(PermSecurityWrite), s.palDefenderRESTToken)
	api.POST("/security/paldefender/reload-config", Require(PermSecurityWrite), s.palDefenderReloadConfig)
	api.GET("/security/paldefender/gm/status", s.palDefenderGMStatus)
	api.GET("/security/paldefender/gm/players", s.palDefenderGMPlayers)
	api.GET("/security/paldefender/gm/items", s.palDefenderGMItems)
	api.GET("/security/paldefender/gm/players/:id", s.palDefenderGMPlayer)
	api.GET("/security/paldefender/gm/players/:id/inventory", s.palDefenderGMInventory)
	api.POST("/security/paldefender/gm/players/:id/items", Require(PermPlayersWrite), s.palDefenderGMGiveItems)
	api.POST("/security/paldefender/gm/players/:id/items/remove", Require(PermPlayersWrite), s.palDefenderGMRemoveItems)
	api.POST("/security/paldefender/gm/players/:id/teleport", Require(PermPlayersWrite), s.palDefenderGMTeleport)
	api.GET("/security/paldefender/gm/players/:id/progression", s.palDefenderGMProgression)
	api.POST("/security/paldefender/gm/players/:id/progression", Require(PermPlayersWrite), s.palDefenderGMGiveProgression)
	api.GET("/security/paldefender/gm/players/:id/techs", s.palDefenderGMTechs)
	api.POST("/security/paldefender/gm/players/:id/techs/learn", Require(PermPlayersWrite), s.palDefenderGMLearnTech)
	api.POST("/security/paldefender/gm/players/:id/techs/forget", Require(PermPlayersWrite), s.palDefenderGMForgetTech)
	api.GET("/security/paldefender/gm/players/:id/pals", s.palDefenderGMPals)
	api.POST("/security/paldefender/gm/players/:id/pals", Require(PermPlayersWrite), s.palDefenderGMGivePals)
	api.POST("/security/paldefender/gm/players/:id/custom-pal", Require(PermPlayersWrite), s.palDefenderGMGiveCustomPal)
	api.POST("/security/paldefender/gm/players/:id/custom-pals", Require(PermPlayersWrite), s.palDefenderGMGiveCustomPals)
	api.POST("/security/paldefender/gm/players/:id/pals/release", Require(PermPlayersWrite), s.palDefenderGMReleasePal)
	api.POST("/security/paldefender/gm/players/:id/pal-templates", Require(PermPlayersWrite), s.palDefenderGMGivePalTemplates)
	api.POST("/security/paldefender/gm/players/:id/export-pals", Require(PermPlayersWrite), s.palDefenderGMExportPals)
	api.GET("/security/paldefender/gm/players/:id/exported-pal-templates", s.palDefenderGMExportedPalTemplates)
	api.GET("/security/paldefender/gm/players/:id/exported-pal-templates/:name", s.palDefenderGMExportedPalTemplate)
	api.POST("/security/paldefender/gm/players/:id/message", Require(PermPlayersWrite), s.palDefenderGMSendMessage)
	api.POST("/security/paldefender/gm/players/:id/kick", Require(PermPlayersWrite), s.palDefenderGMKick)
	api.POST("/security/paldefender/gm/players/:id/ban", Require(PermPlayersWrite), s.palDefenderGMBan)
	api.POST("/security/paldefender/gm/players/:id/unban", Require(PermPlayersWrite), s.palDefenderGMUnban)
	api.POST("/security/paldefender/gm/broadcast", Require(PermPlayersWrite), s.palDefenderGMBroadcast)
	api.GET("/security/paldefender/gm/commands", s.palDefenderGMCommandCatalog)
	api.GET("/security/paldefender/gm/commands/runtime", Require(PermSecurityWrite), s.palDefenderGMRCONCommands)
	api.GET("/security/paldefender/gm/catalog/technology", s.palDefenderGMTechnologyCatalog)
	api.GET("/security/paldefender/gm/catalog/technologies", s.palDefenderGMLocalTechnologyCatalog)
	api.GET("/security/paldefender/gm/catalog/pals", s.palDefenderGMPalCatalog)
	api.GET("/security/paldefender/gm/catalog/passives", s.palDefenderGMPassiveCatalog)
	api.GET("/security/paldefender/gm/catalog/skins", s.palDefenderGMSkinCatalog)
	api.GET("/security/paldefender/gm/catalog/references", s.palDefenderGMReferences)
	api.GET("/security/paldefender/gm/pal-templates", s.palDefenderGMListTemplates)
	api.GET("/security/paldefender/gm/pal-templates/:name", s.palDefenderGMGetTemplate)
	api.PUT("/security/paldefender/gm/pal-templates/:name", Require(PermSecurityWrite), s.palDefenderGMPutTemplate)
	api.DELETE("/security/paldefender/gm/pal-templates/:name", Require(PermSecurityWrite), s.palDefenderGMDeleteTemplate)
	api.GET("/security/paldefender/starter-gift", s.starterGiftConfig)
	api.PUT("/security/paldefender/starter-gift", Require(PermSecurityWrite), s.putStarterGiftConfig)
	api.POST("/security/paldefender/starter-gift/grants/:id/retry", Require(PermSecurityWrite), s.retryStarterGift)
	api.DELETE("/security/paldefender/starter-gift/grants/:id", Require(PermSecurityWrite), s.forgetStarterGift)
	api.GET("/security/paldefender/access", Require(PermSecurityWrite), s.palDefenderAccessSettings)
	api.PUT("/security/paldefender/access", Require(PermSecurityWrite), s.palDefenderPutAccessSettings)
	api.GET("/security/paldefender/whitelist", Require(PermSecurityWrite), s.palDefenderWhitelist)
	api.POST("/security/paldefender/whitelist/:id", Require(PermSecurityWrite), s.palDefenderWhitelistAdd)
	api.DELETE("/security/paldefender/whitelist/:id", Require(PermSecurityWrite), s.palDefenderWhitelistRemove)
	api.POST("/security/paldefender/admins/:id/toggle", Require(PermSecurityWrite), s.palDefenderSetAdmin)
}

func (s Server) registerWorldRoutes(api *gin.RouterGroup) {
	api.GET("/save-sources", s.listSaveSources)
	api.POST("/save-sources/import/inspect", Require(PermServerControl), s.inspectSaveSourceImportDispatch)
	api.POST("/save-sources/import/inspect/:id/select", Require(PermServerControl), s.selectSaveSourceImportCandidate)
	api.POST("/save-sources/import", Require(PermServerControl), s.importSaveSource)
	api.POST("/save-sources/:id/activate", Require(PermServerControl), s.activateSaveSource)
	api.POST("/save-sources/:id/rebuild", Require(PermServerControl), s.rebuildSaveSource)
	api.PATCH("/save-sources/:id", Require(PermServerControl), s.renameSaveSource)
	api.DELETE("/save-sources/:id", Require(PermServerControl), s.deleteSaveSource)
	api.GET("/save/index/status", s.saveIndexStatus)
	api.POST("/save/index/rebuild", Require(PermServerControl), s.saveIndexRebuild)
	api.GET("/players", s.listSavePlayers)
	api.GET("/guilds", s.listSaveGuilds)
	api.GET("/guilds/:id", s.getSaveGuild)
	api.GET("/bases", s.listSaveBases)
	api.GET("/bases/:id", s.getSaveBase)
	api.PUT("/bases/:id/name", Require(PermServerControl), s.putBaseCustomName)
	api.DELETE("/bases/:id/name", Require(PermServerControl), s.deleteBaseCustomName)
	api.GET("/bases/:id/storage", s.getSaveBaseStorage)
	api.GET("/bases/:id/workers", s.getSaveBaseWorkers)
	api.GET("/bases/:id/feed-boxes", s.getSaveBaseFeedBoxes)
	api.GET("/inventory", s.listGlobalInventory)
	api.GET("/pals", s.listSavePals)
	api.GET("/pals/:id", s.getSavePal)
	api.GET("/map/entities", s.listMapEntities)
	api.GET("/breeding/catalog", s.breedingCatalog)
	api.GET("/breeding/status", s.breedingStatus)
	api.GET("/breeding/history", s.breedingHistory)
	api.GET("/breeding/presets", s.listBreedingPresets)
	api.POST("/breeding/presets", Require(PermRead), s.putBreedingPreset)
	api.DELETE("/breeding/presets/:id", Require(PermRead), s.deleteBreedingPreset)
	api.GET("/breeding/custom-containers", s.listCustomPalContainers)
	api.POST("/breeding/custom-containers", Require(PermRead), s.putCustomPalContainer)
	api.DELETE("/breeding/custom-containers/:id", Require(PermRead), s.deleteCustomPalContainer)
	api.POST("/breeding/jobs", Require(PermRead), s.submitBreedingJob)
	api.GET("/breeding/jobs/:id/result", s.breedingResult)
	api.POST("/breeding/jobs/:id/pause", Require(PermRead), s.controlBreedingJob("pause"))
	api.POST("/breeding/jobs/:id/resume", Require(PermRead), s.controlBreedingJob("resume"))
	api.POST("/breeding/jobs/:id/cancel", Require(PermRead), s.controlBreedingJob("cancel"))
	api.GET("/players/bans", s.listPlayerBans)
	api.POST("/players/bans", Require(PermPlayersWrite), s.addPlayerBan)
	api.DELETE("/players/bans/:steam_id", Require(PermPlayersWrite), s.deletePlayerBan)
	api.GET("/players/whitelist", s.listPlayerWhitelist)
	api.PUT("/players/whitelist", Require(PermPlayersWrite), s.putPlayerWhitelist)
	api.POST("/players/:id/kick", Require(PermPlayersWrite), s.kickPlayer)
	api.POST("/players/:id/ban", Require(PermPlayersWrite), s.banPlayer)
	api.POST("/players/:id/unban", Require(PermPlayersWrite), s.unbanPlayer)
	api.GET("/players/:id/inventory", s.getSavePlayerInventory)
	api.PUT("/players/:id/annotation", Require(PermPlayersWrite), s.putPlayerAnnotation)
	api.DELETE("/players/:id/annotation", Require(PermPlayersWrite), s.deletePlayerAnnotation)
	api.GET("/players/:id", s.getSavePlayer)
}

func (s Server) registerFrontendRoutes(router *gin.Engine) {
	registerFrontendFilesystem(router, s.webUI)
}

func registerFrontendFilesystem(router *gin.Engine, files fs.FS) {
	serve := func(name, cacheControl string) gin.HandlerFunc {
		return func(c *gin.Context) {
			body, err := fs.ReadFile(files, name)
			if err != nil {
				c.Status(http.StatusNotFound)
				return
			}
			c.Header("Cache-Control", cacheControl)
			http.ServeContent(c.Writer, c.Request, name, time.Time{}, bytes.NewReader(body))
		}
	}
	available := files != nil
	if available {
		index := serve("index.html", "no-cache")
		router.GET("/", index)
		router.HEAD("/", index)
		asset := func(c *gin.Context) {
			requested := strings.TrimPrefix(c.Param("path"), "/")
			name := "assets/" + requested
			if requested == "" || !fs.ValidPath(name) {
				c.Status(http.StatusNotFound)
				return
			}
			serve(name, "public, max-age=31536000, immutable")(c)
		}
		router.GET("/assets/*path", asset)
		router.HEAD("/assets/*path", asset)
		if info, err := fs.Stat(files, "favicon.ico"); err == nil && !info.IsDir() {
			favicon := serve("favicon.ico", "public, max-age=3600")
			router.GET("/favicon.ico", favicon)
			router.HEAD("/favicon.ico", favicon)
		}
	}
	router.NoRoute(func(c *gin.Context) {
		if c.Request.URL.Path == "/api" || strings.HasPrefix(c.Request.URL.Path, "/api/") {
			fail(c, http.StatusNotFound, "not_found", "api route not found")
			return
		}
		if !available || (c.Request.Method != http.MethodGet && c.Request.Method != http.MethodHead) {
			c.Status(http.StatusNotFound)
			return
		}
		requested := strings.TrimPrefix(c.Request.URL.Path, "/")
		if requested != "" && fs.ValidPath(requested) {
			if info, err := fs.Stat(files, requested); err == nil && !info.IsDir() {
				serve(requested, "public, max-age=3600")(c)
				return
			}
		}
		serve("index.html", "no-cache")(c)
	})
}