# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PalPanel API
  version: 1.0.0
  description: Stable HTTP contract for the PalPanel 1.x management API.
servers:
  - url: /api
security:
  - sessionCookie: []
  - developmentKey: []
...
paths:
  /community-servers:
    get:
      operationId: listCommunityServers
      tags: [server]
      x-palpanel-permission: read
      parameters: &communityServerQueryParameters
        - {name: region, in: query, schema: {type: string, enum: [cn, global], default: cn}}
        - {name: search, in: query, schema: {type: string, maxLength: 128}}
        - {name: min_players, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: max_players, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: password, in: query, schema: {type: boolean}}
        - {name: version, in: query, schema: {type: string, maxLength: 64}}
        - {name: status, in: query, schema: {type: string, enum: [online, offline, all], default: online}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 30}}
      responses: {'200': *communityServerResult, '400': *error, '401': *error, '429': *error, '502': *error, '503': *error}
  /community-servers/source-status:
    get: {operationId: getCommunityServerSourceStatus, tags: [server], x-palpanel-permission: read, responses: {'200': *communityServerStatus, '401': *error}}
  /community-servers/refresh:
    post:
      operationId: refreshCommunityServers
      tags: [server]
      x-palpanel-permission: read
      parameters: *communityServerQueryParameters
      responses: {'200': *communityServerResult, '400': *error, '401': *error, '403': *error, '429': *error, '502': *error, '503': *error}
...
  /auth/status:
    get:
      operationId: getAuthenticationStatus
      tags: [auth]
      security: []
      x-palpanel-permission: public
      responses: {'200': *authStatus, '500': *error}
  /auth/register:
    post:
      operationId: registerInitialAdministrator
      tags: [auth]
      security: []
      x-palpanel-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthCredentials'}
      responses: {'201': *session, '400': *error, '409': *error, '422': *error, '429': *error, '500': *error}
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      x-palpanel-permission: public
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthCredentials'}
      responses: {'200': *session, '400': *error, '401': *error, '429': *error, '500': *error}
  /auth/me:
    get: {operationId: getCurrentPrincipal, tags: [auth], x-palpanel-permission: read, responses: {'200': *session, '401': *error}}
  /auth/logout:
    post: {operationId: logout, tags: [auth], x-palpanel-permission: read, responses: *writeResponses}
  /auth/password:
    put:
      operationId: changeOwnPassword
      tags: [auth]
      security: [{sessionCookie: []}]
      x-palpanel-permission: session
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PasswordChangeRequest'}
      responses: {'200': *success, '400': *error, '401': *error, '403': *error, '422': *error, '429': *error, '500': *error}
  /auth/api-keys:
    get: {operationId: listDevelopmentKeys, tags: [auth], x-palpanel-permission: read, responses: {'200': *developmentKeys, '401': *error, '500': *error}}
    post:
      operationId: createDevelopmentKey
      tags: [auth]
      x-palpanel-permission: 'security:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DevelopmentKeyInput'}
      responses: {'201': *developmentKey, '400': *error, '401': *error, '403': *error, '422': *error, '500': *error}
  /auth/api-keys/{id}:
    parameters: [*idParameter]
    delete: {operationId: revokeDevelopmentKey, tags: [auth], x-palpanel-permission: 'security:write', responses: *writeResponses}
  /jobs:
    get: {operationId: listJobs, tags: [system], x-palpanel-permission: read, responses: *readResponses}
  /jobs/{id}:
    parameters: [*idParameter]
    get: {operationId: getJob, tags: [system], x-palpanel-permission: read, responses: *readResponses}
  /patch/update/status:
    get: {operationId: getPatchUpdateStatus, tags: [system], x-palpanel-permission: read, responses: {'200': *success, '401': *error, '502': *error}}
  /patch/update/check:
    post: {operationId: checkPatchUpdate, tags: [system], x-palpanel-permission: 'server:control', responses: {'202': *accepted, '401': *error, '403': *error, '500': *error}}
  /patch/update:
    post: {operationId: applyPatchUpdate, tags: [system], x-palpanel-permission: 'server:control', responses: {'202': *accepted, '401': *error, '403': *error, '500': *error}}
  /audit-logs:
    get: {operationId: listAuditLogs, tags: [system], x-palpanel-permission: 'audit:read', responses: *readResponses}
  /system/debug:
    get: {operationId: getDebugLogging, tags: [system], x-palpanel-permission: read, responses: *readResponses}
    put: {operationId: updateDebugLogging, tags: [system], x-palpanel-permission: 'config:write', requestBody: *jsonRequest, responses: *writeResponses}
  /settings/network-proxy:
    get: {operationId: getNetworkProxyConfig, tags: [system], x-palpanel-permission: read, responses: {'200': *networkProxyConfig, '401': *error, '500': *error}}
    put:
      operationId: updateNetworkProxyConfig
      tags: [system]
      x-palpanel-permission: 'config:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NetworkProxyConfigUpdate'}
      responses: {'200': *networkProxyConfig, '400': *error, '401': *error, '403': *error, '500': *error}
  /settings/network-proxy/test:
    post:
      operationId: testNetworkProxy
      tags: [system]
      x-palpanel-permission: 'config:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NetworkProxyTestRequest'}
      responses: {'200': *networkProxyTest, '400': *error, '401': *error, '403': *error, '502': *error}
  /alerts:
    get: {operationId: listAlerts, tags: [system], x-palpanel-permission: read, responses: *readResponses}
  /alerts/{id}/ack:
    parameters: [*idParameter]
    post: {operationId: acknowledgeAlert, tags: [system], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /schedules:
    get: {operationId: listSchedules, tags: [system], x-palpanel-permission: read, responses: *readResponses}
    post:
      operationId: createSchedule
      tags: [system]
      x-palpanel-permission: 'server:control'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScheduleInput'}
      responses: *writeResponses
  /schedules/{id}:
    parameters: [*idParameter]
    put:
      operationId: updateSchedule
      tags: [system]
      x-palpanel-permission: 'server:control'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScheduleInput'}
      responses: *writeResponses
    delete: {operationId: deleteSchedule, tags: [system], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /schedules/{id}/run:
    parameters: [*idParameter]
    post: {operationId: runSchedule, tags: [system], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/status:
    get: {operationId: getServerStatus, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/prerequisites:
    get: {operationId: getServerPrerequisites, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/host:
    get: {operationId: getHostCapabilities, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/runtime:
    get: {operationId: getRuntime, tags: [server], x-palpanel-permission: read, responses: *readResponses}
    put: {operationId: updateRuntime, tags: [server], x-palpanel-permission: 'config:write', requestBody: *jsonRequest, responses: *writeResponses}
  /server/import:
    post: {operationId: importServerDirectory, tags: [server], x-palpanel-permission: 'config:write', requestBody: *jsonRequest, responses: *writeResponses}
  /server/docker/plan:
    get: {operationId: getDockerInstallPlan, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/docker/install:
    post: {operationId: installDocker, tags: [server], x-palpanel-permission: 'server:control', requestBody: *jsonRequest, responses: *writeResponses}
  /server/docker/mirrors/plan:
    get: {operationId: getDockerMirrorPlan, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/docker/mirrors/configure:
    post: {operationId: configureDockerMirrors, tags: [server], x-palpanel-permission: 'server:control', requestBody: *jsonRequest, responses: *writeResponses}
  /server/bootstrap:
    post: {operationId: bootstrapServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/logs:
    get:
      operationId: getServerLogs
      tags: [server]
      x-palpanel-permission: read
      parameters:
        - {name: channel, in: query, required: false, schema: {type: string, enum: [game, launcher, paldefender-rest]}}
        - {name: tail, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 5000, default: 200}}
        - {name: search, in: query, required: false, schema: {type: string}}
        - {name: level, in: query, required: false, schema: {type: string}}
        - {name: since, in: query, required: false, schema: {type: string}}
      responses: *readResponses
  /server/world:
    get: {operationId: getWorld, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /server/world/reset:
    post: {operationId: resetWorld, tags: [server], x-palpanel-permission: 'world:reset', requestBody: *jsonRequest, responses: *writeResponses}
  /server/install:
    post: {operationId: installServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/update:
    post: {operationId: updateServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/update-if-needed:
    post: {operationId: updateServerIfNeeded, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
...
  /server/version/check:
    post: {operationId: checkServerVersion, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/start:
    post: {operationId: startServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/stop:
    post: {operationId: stopServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/restart:
    post: {operationId: restartServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/safe-restart:
    post:
      operationId: safeRestartServer
      tags: [server]
      x-palpanel-permission: 'server:control'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SafeLifecycleRequest'}}}}
      responses: *writeResponses
  /server/safe-stop:
    post:
      operationId: safeStopServer
      tags: [server]
      x-palpanel-permission: 'server:control'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SafeLifecycleRequest'}}}}
      responses: *writeResponses
  /server/force-stop:
    post: {operationId: forceStopServer, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /server/startup:
    get: {operationId: getServerStartup, tags: [server], x-palpanel-permission: read, responses: *readResponses}
    put: {operationId: updateServerStartup, tags: [server], x-palpanel-permission: 'config:write', requestBody: *jsonRequest, responses: *writeResponses}
  /server/initialize-config:
    post: {operationId: initializeServerConfig, tags: [server], x-palpanel-permission: 'config:write', responses: *writeResponses}
  /monitor/snapshot:
    get: {operationId: getMonitorSnapshot, tags: [server], x-palpanel-permission: read, responses: {'200': *monitorSnapshot, '401': *error, '500': *error}}
  /monitor/history:
    get: {operationId: getMonitorHistory, tags: [server], x-palpanel-permission: read, responses: {'200': *monitorHistory, '401': *error, '500': *error}}
  /server/backup:
    post: {operationId: createBackup, tags: [server], x-palpanel-permission: 'backup:write', responses: *writeResponses}
  /backups:
    get: {operationId: listBackups, tags: [server], x-palpanel-permission: read, responses: *readResponses}
  /backups/webdav/config:
    get: {operationId: getBackupWebDAVConfig, tags: [server], x-palpanel-permission: 'backup:write', responses: *readResponses}
    put:
      operationId: updateBackupWebDAVConfig
      tags: [server]
      x-palpanel-permission: 'backup:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WebDAVConfigUpdate'}
      responses: *writeResponses
  /backups/webdav/test:
    post:
      operationId: testBackupWebDAVConfig
      tags: [server]
      x-palpanel-permission: 'backup:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WebDAVConfigUpdate'}
      responses: *writeResponses
  /backups/{name}/restore:
    parameters: [*nameParameter]
    post: {operationId: restoreBackup, tags: [server], x-palpanel-permission: 'backup:write', responses: *writeResponses}
  /backups/{name}/download:
    parameters: [*nameParameter]
    get: {operationId: downloadBackup, tags: [server], x-palpanel-permission: 'backup:write', responses: *readResponses}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: palpanel_session
      description: HttpOnly SameSite=Lax session cookie created by registration or login.
    developmentKey:
      type: http
      scheme: bearer
      bearerFormat: ppk_...
      description: 'Revocable development key supplied as Authorization: Bearer ppk_...'
  schemas:
...
    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 32}
        password: {type: string, format: password, minLength: 12, maxLength: 128}
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password: {type: string, format: password, minLength: 1, maxLength: 128}
        new_password: {type: string, format: password, minLength: 12, maxLength: 128}
...
    DevelopmentKeyInput:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
...
    NetworkProxyTestRequest:
      type: object
      additionalProperties: false
      required: [scope]
      properties:
        scope: {type: string, enum: [install, community]}
...